pyproject.toml
[project]
name = "france-jobs"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "playwright",
    "beautifulsoup4",
    "httpx",
    "python-dotenv",
    "openpyxl",
    "pandas",
    "markdownify",
]

